RAG-Q&A-CHATBOT-WITH-PDF-&-CHAT-HISTORY/requirements.txt
streamlit
langchain
langchain-chroma
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
pypdf
python-dotenv
sentence-transformers